Download PDFs

Takes the results of an arxiv API query and downloads papers to disk by default. Saving to an AWS S3 bucket is also supported, though credentials must be set as environment variables prior to using download_pdf (see example). Papers are saved with their arXiv id, e.g. "2001.12345v1.pdf"

download_pdf(data, link_col = "link_pdf", directory = ".", bucket = NULL)

Arguments

data
data frame of arXiv records

link_col
name of column containing the PDF locations. link_pdf by default.

directory
directory to download files to. Current working directory by default.

bucket
name of AWS S3 bucket to save files to.

Value

None

Examples

# get paper metadata, then download it to working directory
# ml <- get_records("cs.LG", "20200801 TO 20200802", 1)
# download_pdf(ml)

# set up S3 credentials, then download to bucket
# Sys.setenv("AWS_ACCESS_KEY_ID" = "key")
# Sys.setenv("AWS_SECRET_ACCESS_KEY" = "secretkey")
# Sys.setenv("AWS_DEFAULT_REGION" = "us-east-2")
# download_pdf(ml, bucket="mybucket")